+++ src/protocols/gamespy/protocols/one/protocol.rs
@@ -226,7 +226,7 @@ pub fn query(address: &SocketAddr, timeout_settings: Option<TimeoutSettings>) ->
             .remove("AdminName")
             .or_else(|| server_vars.remove("admin")),
         has_password: has_password(&mut server_vars)?,
-        game_type: server_vars
+        game_mode: server_vars
             .remove("gametype")
             .ok_or(GDErrorKind::PacketBad)?,
         game_version: server_vars